Move snapshot path builder
