+++ src/test/java/org/apache/cassandra/sidecar/snapshots/PathBuilderTest.java
@@ -1,0 +1,16 @@
+package org.apache.cassandra.sidecar.snapshots;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import io.vertx.core.Vertx;
+import io.vertx.junit5.VertxExtension;
+import org.apache.cassandra.sidecar.cluster.InstancesConfig;
+
+@ExtendWith(VertxExtension.class)
+class PathBuilderTest extends AbstractPathBuilderTest
+{
+    PathBuilder initialize(Vertx vertx, InstancesConfig instancesConfig)
+    {
+        return new PathBuilder(vertx.fileSystem(), instancesConfig);
+    }
+}
