--- src/test/java/org/apache/cassandra/sidecar/utils/SnapshotPathBuilderTest.java
@@ -1,16 +1,0 @@
-package org.apache.cassandra.sidecar.utils;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import io.vertx.core.Vertx;
-import io.vertx.junit5.VertxExtension;
-import org.apache.cassandra.sidecar.cluster.InstancesConfig;
-
-@ExtendWith(VertxExtension.class)
-class SnapshotPathBuilderTest extends AbstractSnapshotPathBuilderTest
-{
-    SnapshotPathBuilder initialize(Vertx vertx, InstancesConfig instancesConfig)
-    {
-        return new SnapshotPathBuilder(vertx.fileSystem(), instancesConfig);
-    }
-}
